# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ord",
    "description": "◉ Ordinal wallet and block explorer",
    "license": {
      "name": "CC0-1.0"
    },
    "version": "0.14.1"
  },
  "paths": {
...
    "/api/v1/brc20/address/{address}/transferable": {
      "get": {
        "tags": [
          "brc20"
        ],
        "summary": "Get the balance of ticker of the address.",
        "description": "Get the balance of ticker of the address.\n\nRetrieve the balance of the ticker from the given address.",
        "operationId": "brc20_all_transferable",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "description": "Address",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Obtain account all transferable inscriptions.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiBRC20Transferable"
                }
              }
            }
          },
          "400": {
            "description": "Bad query.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 2,
                  "msg": "bad request"
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 3,
                  "msg": "not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 1,
                  "msg": "internal error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/brc20/tick": {
      "get": {
        "tags": [
          "brc20"
        ],
        "summary": "Get all tickers info.",
        "description": "Get all tickers info.\n\nRetrieve detailed information about all tickers.",
        "operationId": "brc20_all_tick_info",
        "responses": {
          "200": {
            "description": "Obtain matching all BRC20 tickers.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiBRC20AllTick"
                }
              }
            }
          },
          "400": {
            "description": "Bad query.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 2,
                  "msg": "bad request"
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 3,
                  "msg": "not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 1,
                  "msg": "internal error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/brc20/tick/{ticker}": {
      "get": {
        "tags": [
          "brc20"
        ],
        "summary": "Get the ticker info.",
        "description": "Get the ticker info.\n\nRetrieve detailed information about the ticker.",
        "operationId": "brc20_tick_info",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "description": "Token ticker",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 4,
              "minLength": 4
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Obtain matching BRC20 ticker by query.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiBRC20Tick"
                }
              }
            }
          },
          "400": {
            "description": "Bad query.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 2,
                  "msg": "bad request"
                }
              }
            }
          },
          "404": {
            "description": "Ticker not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 3,
                  "msg": "not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 1,
                  "msg": "internal error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/brc20/tick/{ticker}/address/{address}/transferable": {
      "get": {
        "tags": [
          "brc20"
        ],
        "summary": "Get the transferable inscriptions of the address.",
        "description": "Get the transferable inscriptions of the address.\n\nRetrieve the transferable inscriptions with the ticker from the given address.",
        "operationId": "brc20_transferable",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "description": "Token ticker",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 4,
              "minLength": 4
            }
          },
          {
            "name": "address",
            "in": "path",
            "description": "Address",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Obtain account transferable inscriptions of ticker.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiBRC20Transferable"
                }
              }
            }
          },
          "400": {
            "description": "Bad query.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 2,
                  "msg": "bad request"
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 3,
                  "msg": "not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 1,
                  "msg": "internal error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiBRC20AllTick": {
        "type": "object",
        "required": [
          "code",
          "msg",
          "data"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "$ref": "#/components/schemas/brc20.ApiTickInfos"
          },
          "msg": {
            "type": "string",
            "description": "ok",
            "example": "ok"
          }
        }
      },
...
      "ApiBRC20Tick": {
        "type": "object",
        "required": [
          "code",
          "msg",
          "data"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "$ref": "#/components/schemas/brc20.ApiTickInfo"
          },
          "msg": {
            "type": "string",
            "description": "ok",
            "example": "ok"
          }
        }
      },
      "ApiBRC20Transferable": {
        "type": "object",
        "required": [
          "code",
          "msg",
          "data"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "$ref": "#/components/schemas/brc20.ApiTransferableAssets"
          },
          "msg": {
            "type": "string",
            "description": "ok",
            "example": "ok"
          }
        }
      },
...
      "ApiError": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "Internal"
            ],
            "properties": {
              "Internal": {
                "type": "string",
                "description": "Internal server error."
              }
            },
            "example": {
              "code": 1,
              "msg": "internal error"
            }
          },
          {
            "type": "object",
            "required": [
              "BadRequest"
            ],
            "properties": {
              "BadRequest": {
                "type": "string",
                "description": "Bad request."
              }
            },
            "example": {
              "code": 1,
              "msg": "bad request"
            }
          },
          {
            "type": "object",
            "required": [
              "NotFound"
            ],
            "properties": {
              "NotFound": {
                "type": "string",
                "description": "Resource not found."
              }
            },
            "example": {
              "code": 1,
              "msg": "not found"
            }
          }
        ]
      },
...
      "ScriptPubkey": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "address"
            ],
            "properties": {
              "address": {
                "type": "string",
                "description": "Address."
              }
            }
          },
          {
            "type": "object",
            "required": [
              "nonStandard"
            ],
            "properties": {
              "nonStandard": {
                "type": "string",
                "description": "Non-standard script hash."
              }
            }
          }
        ]
      },
...
      "brc20.ApiTickInfo": {
        "type": "object",
        "description": "Description of a BRC20 ticker.",
        "required": [
          "tick",
          "inscriptionId",
          "inscriptionNumber",
          "supply",
          "burnedSupply",
          "selfMint",
          "limitPerMint",
          "minted",
          "decimal",
          "deployBy",
          "txid",
          "deployHeight",
          "deployBlocktime"
        ],
        "properties": {
          "burnedSupply": {
            "type": "string",
            "format": "uint64",
            "description": "The amount of the ticker that has been burned."
          },
          "decimal": {
            "type": "integer",
            "format": "uint8",
            "description": "The decimal of the ticker.<br>\nNumber of decimals cannot exceed 18 (default).",
            "default": 18,
            "example": 18,
            "maximum": 18,
            "minimum": 0
          },
          "deployBlocktime": {
            "type": "integer",
            "format": "uint32",
            "description": "The timestamp of the block that the ticker deployed.",
            "minimum": 0
          },
          "deployBy": {
            "$ref": "#/components/schemas/ScriptPubkey"
          },
          "deployHeight": {
            "type": "integer",
            "format": "uint32",
            "description": "The height of the block that the ticker deployed.",
            "minimum": 0
          },
          "inscriptionId": {
            "type": "string",
            "description": "Inscription ID of the ticker deployed."
          },
          "inscriptionNumber": {
            "type": "integer",
            "format": "int32",
            "description": "Inscription number of the ticker deployed."
          },
          "limitPerMint": {
            "type": "string",
            "format": "uint64",
            "description": "The maximum amount of each mining."
          },
          "minted": {
            "type": "string",
            "format": "uint64",
            "description": "The amount of the ticker that has been minted."
          },
          "selfMint": {
            "type": "boolean",
            "description": "Whether the ticker is self minted."
          },
          "supply": {
            "type": "string",
            "format": "uint64",
            "description": "The total supply of the ticker.<br>\nMaximum supply cannot exceed uint64_max.\n\nA string containing a 64-bit unsigned integer.<br>\nWe represent u64 values as a string to ensure compatibility with languages such as JavaScript that do not parse u64s in JSON natively."
          },
          "tick": {
            "type": "string",
            "description": "Name of the ticker."
          },
          "txid": {
            "type": "string",
            "description": "A hex encoded 32 byte transaction ID that the ticker deployed.\n\nThis is represented in a string as adding a prefix 0x to a 64 character hex string."
          }
        }
      },
      "brc20.ApiTickInfos": {
        "type": "object",
        "required": [
          "tokens"
        ],
        "properties": {
          "tokens": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/brc20.ApiTickInfo"
            }
          }
        }
      },
...
      "brc20.ApiTransferableAsset": {
        "type": "object",
        "required": [
          "inscriptionId",
          "inscriptionNumber",
          "amount",
          "tick",
          "owner",
          "location"
        ],
        "properties": {
          "amount": {
            "type": "string",
            "format": "uint64",
            "description": "The amount of the ticker that will be transferred."
          },
          "inscriptionId": {
            "type": "string",
            "description": "The inscription id."
          },
          "inscriptionNumber": {
            "type": "integer",
            "format": "int32",
            "description": "The inscription number."
          },
          "location": {
            "type": "string",
            "description": "The inscription location."
          },
          "owner": {
            "type": "string",
            "description": "The address to which the transfer will be made."
          },
          "tick": {
            "type": "string",
            "description": "The ticker name that will be transferred."
          }
        }
      },
      "brc20.ApiTransferableAssets": {
        "type": "object",
        "required": [
          "inscriptions"
        ],
        "properties": {
          "inscriptions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/brc20.ApiTransferableAsset"
            }
          }
        }
      },
...
    }
  }
}